chore: move `Task` and related types to separate file
